refactor: 재입력 로직 수정
변경 내용 :
매 입력마다 시도 횟수가 제한되도록 수정
패키지 구조 변경
