+++ src/main/java/chess/controller/TryCount.java
@@ -1,18 +1,23 @@
-package chess.controller.action;
+package chess.controller;
 
 public class TryCount {
 
     private static final int COUNT_UNIT = 1;
+    private static final int INITIAL_COUNT = 0;
     private final int maximumTry;
     private int tryCount;
 
     public TryCount(final int maximumTry) {
         this.maximumTry = maximumTry;
-        tryCount = 0;
+        tryCount = INITIAL_COUNT;
     }
 
     public boolean canRetry() {
         return maximumTry > tryCount;
+    }
+
+    public void resetCount() {
+        tryCount = 0;
     }
 
     public void count() {
